Rename post to follow _posts/ convention
Local build was failing to include _posts/ and reason may be that
the only file I had in the dir didn't follow the file naming
convention YYYY-MM-DD-name-of-post.ext
So it looked like _posts/ was empty and not included.
Changed file name to follow convention and will follow convention
for all future additions.