Changed mp3s/ to podcasts/
RSS feed included inconsistent references to the same folder -
some were to mp3s/ others to podcasts/

Changed folder to podcasts/ and changed all references in feed.xml
accordingly. Also fixed some indentations in feed.xml.
This fix seems to have sorted out the item loop in the RSS feed,
which is now showing episode 1 in the feed, including the episode
duration and filesize which were calculated using the Ruby plugin
audioinfo. So this is working locally at least.