Fix mp3 url and remove categories in _posts/ ep 1
Pointed mp3 url to where the audio file currently lives.
Need to update this to use a Podtrac redirect once that's
set up.

Also commented out the categories listed in the frontmatter,
because these were being included in the post slug leading
to a massive url. Podcast categories are better served in the
RSS feed anyway.