Renamed site.url to site.prod_url
I was using the site.url tag to pass the website's root to various
RSS tags. But I have had to rename the url tag to something else,
and chosen prod_url. The url tag is a default variable that exists
in Jekyll. Usually url can set to anything else, if the site gets
built by GitHub Pages. But because my site contains custom Ruby
plugins, I need to build the site in a development environment
and then push the build to my gh-pages branch. In a development
environment, site.url is automatically set to the value of host,
port, and SSL-related options, i.e. http://localhost:4000
Obviously this makes the built RSS point to an erroneous url.
See https://jekyllrb.com/docs/variables/ for more info.