Fixed embedded audioplayer and its include tag in posts
Various changes to fix the audioplayer embed and the include tag
in posts to include the audioplayer. The main error I was getting
was because  I was trying to pass {{ post.mp3_url }} as a
parameter in the include. But curly brackets inside includes
statements cause an error. So I had to {% assign %} post.mp3_url
to a variable first then pass that to include. Various other
variable name changes included in this fix, including a fix to
the _includes/ dir (following convention) which I'd previously
named _include/ by mistake.