Fixed audio player and added css for audio
The audio player wasn't working because I wasn't properly
passing it the URL to the mp3 file. It being a post, I thought
that the parent name for YML variables was post,
e.g. post.mp3_url. But it's actually page.mp3_url. While I was at
it I tidied up the audioplayer html and added a chunk of css style
before the post content, mainly to make the audio player much wider.