Remove <itunes:block> tag in RSS feed
Removed the block tag, which to allow the podcaster to block the
feed from getting updated. The feed I'd used included the block
tag set to "yes", blocking itunes updates. This is hopefully also
the reason why Google is not picking up my feed. The itunes
documentation said that block==yes stops feed update and any other
value will allow updates. But I'd tried setting block==no and
itunes is still saying my feed is blocked by user. So I'll try
removing the <block> tag entirely (not required) and seeing if
this causes the feed to go live on itunes and Google.