Add podtrac redirect to mp3 url for analytics
Podtrac is a free service to collect data on number of listens and
downloads my episodes get. Just needed to put a Podtrac prefix in the
url to my podcast episodes and it then points to Podtrac which then
redirects to the episode. Added podtrac_url variable in post YML
frontmatter, which I've used in the episode audioplayer and in the
<enclosure> tag in the RSS feed.